backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
Werkzeug==2.3.7
bcrypt==4.0.1
PyJWT==2.8.0
pymongo==4.6.0
reportlab==4.0.5
weasyprint==60.2
